refine win casting

+++ src/process/luaosutils_process_os_win.cpp
@@ -140,12 +140,12 @@ bool process_launch(const std::string& cmd, const std::string& dir)
 
 void run_event_loop(double timeoutSeconds)
 {
-   ULONGLONG waitTime = std::lround(timeoutSeconds * 1000.0);
-   ULONGLONG elapsed = 0;
+   DWORD waitTime = std::lround(timeoutSeconds * 1000.0);
+   DWORD elapsed = 0;
    MSG msg{};
    DWORD result = 0;
    ULONGLONG start = GetTickCount64();
-   while ((result = ::MsgWaitForMultipleObjects(0, NULL, FALSE, static_cast<DWORD>(waitTime - elapsed), QS_ALLINPUT)) != WAIT_FAILED)
+   while ((result = ::MsgWaitForMultipleObjects(0, NULL, FALSE, waitTime - elapsed, QS_ALLINPUT)) != WAIT_FAILED)
    {
        if (result == WAIT_OBJECT_0)
        {
@@ -155,7 +155,7 @@ void run_event_loop(double timeoutSeconds)
                ::DispatchMessage(&msg);
            }
        }
-       elapsed = GetTickCount64() - start;
+       elapsed = static_cast<DWORD>(GetTickCount64() - start);
        if (elapsed >= waitTime)
            break;
    }
